Expand Exercise feedback to handle unanswered questions
Currently, there's only correct and incorrect feedback. The incorrect
feedback includes unanswered questions too. This commit adds a different
feedback to unanswered questions.

Unanswered questions will be coloured yellow and have the message
'Please select an answer' with the Meh FontAwesome icon ;)

Technical changes:
- Added a new CSS class, unanswered, with the yellow background colour
- Imported faMeh from font awesome
- Change the typing of resultArray state and the temporary
newResultArray variable to include undefined
- resultArray will contain undefined for unanswered quetions, in
addition to the boolean for correct and actually incorrect answers
- getResultClass and getResultMessage will check for undefined given an
index, and if it is undefined, it will render the appropriate CSS class
and message.