Update Exercise-related components to use updated exercise data structure
Previously, exercise data structure was an Object where the key was the
answer and the value was the question. This meant there could only ever
be one question to one answer per exercise set. However, that's not the
reality; there may be multiple questions with the same word as an
answer. New data structure is now a list of Objects, where each Object
has an answer key-value and a question key-value.

We updated the components to reflect the change in the data structure.

Added a new Question interface, and corrected the QuestionsList
interface to actually represent it as an Array.