Components to display subscriptions and bring user to Stripe's Checkout
With the Stripe Subscription extension, Stripe syncs the products
with our Firestore. Using this info we can display the subscriptions.

To bring the client to Stripe Checkout, we first need to create a
checkout session. On creation of this session, a Cloud Function that
Stripe created will get the session ID and put it into Firestore. When
we get the Session ID, we redirect the user to the Checkout page.

The subscribe buttons become disabled once they are clicked. There will
also be a little loading message that appears.

Otherwise, caught errors where possible, which displays a generic error
message and re-enables the subscribe buttons.